requirements.txt
fastapi==0.115.9
Hypercorn==0.17.3
pdf2image==1.17.0
pillow==11.2.1
pytesseract==0.3.13
python-docx==1.1.2
PyPDF2==3.0.1
opencv-python-headless==4.10.0.84
beautifulsoup4==4.13.4
numpy==1.26.4
python-multipart==0.0.9
